Move form schema types to shared folder
