+++ src/packages/shared-types/index.ts
@@ -8,3 +8,4 @@ export * from "./uploads";
 export * from "./actions";
 export * from "./action-types/withdraw-record";
 export * from "./forms";
+export * from "./inputs";
