+++ src/services/ui/src/components/Inputs/calendar.tsx
@@ -1,14 +1,8 @@
 import * as React from "react";
+import { CalendarProps } from "shared-types";
 import { DayPicker } from "react-day-picker";
-
 import { cn } from "@/lib/utils";
 import { buttonVariants } from "./button";
-
-export type CalendarProps = React.ComponentProps<typeof DayPicker> & {
-  className?: string;
-  classNames?: any;
-  showOutsideDays?: boolean;
-};
 
 function Calendar({
   className,
