+++ src/services/ui/src/components/Inputs/switch.tsx
@@ -1,14 +1,8 @@
 /* eslint-disable react/prop-types */
 import * as React from "react";
 import * as SwitchPrimitives from "@radix-ui/react-switch";
-
 import { cn } from "@/lib/utils";
-
-export type SwitchProps = React.ComponentPropsWithoutRef<
-  typeof SwitchPrimitives.Root
-> & {
-  className?: string;
-};
+import { SwitchProps } from "shared-types";
 
 const Switch = React.forwardRef<
   React.ElementRef<typeof SwitchPrimitives.Root> & { className?: string },
