+++ src/services/ui/src/components/Inputs/textarea.tsx
@@ -1,12 +1,6 @@
 import * as React from "react";
-
 import { cn } from "@/lib/utils";
-
-export interface TextareaProps
-  extends React.TextareaHTMLAttributes<HTMLTextAreaElement> {
-  charcount?: "simple" | "limited";
-  charcountstyling?: string;
-}
+import { TextareaProps } from "shared-types";
 
 const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
   ({ className, ...props }, ref) => {
